Mudando classe de ConsultaFaturaController
